Folded everything into one script
